Add bin connector to python.
