+++ sdk/python/src/dagger/connectors/__init__.py
@@ -1,4 +1,4 @@
-from . import docker, http  # noqa
+from . import bin, docker, http  # noqa
 from .base import Config, Connector, get_connector, register_connector
 
 __all__ = [
